final implementation
