+++ patches/query-registry@2.6.0.patch
@@ -1,0 +1,12 @@
+diff --git a/package.json b/package.json
+index 329beade92063613562bc48400cb67f98b5b8c5f..0cc60f07a18de115061e08c55bd9e419f03cd20d 100644
+--- a/package.json
++++ b/package.json
+@@ -4,6 +4,7 @@
+     "version": "2.6.0",
+     "author": "Edoardo Scibona (velut)",
+     "license": "MIT",
++    "type": "module",
+     "source": "./src/index.ts",
+     "main": "./dist/index.js",
+     "module": "./dist/query-registry.esm.js",
