+++ templates/generate-tool-prompt.hbs
