+++ templates/iterate-tool-prompt.hbs
