+++ templates/langchain-tool.hbs
