+++ templates/tool-spec.txt
